requirements_st0ck.txt
# ST0CK Enhanced Requirements
# Combines Alpaca gamma scalping with ST0CK dependencies

# Core Alpaca dependencies
alpaca-py>=0.29.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# Options pricing (for gamma scalping mode)
QuantLib>=1.34

# ST0CK dependencies
aiohttp>=3.8.0
asyncio-throttle>=1.0.0
sqlalchemy>=2.0.0
redis>=4.5.0
flask>=2.3.0
flask-cors>=4.0.0

# Data processing
scipy>=1.10.0
ta>=0.10.0  # Technical indicators

# Logging and monitoring
loguru>=0.7.0
sentry-sdk>=1.0.0

# Time handling
pytz>=2023.3
python-dateutil>=2.8.0

# Development
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
mypy>=1.0.0